Beta Release 0.1! We made it!
This is a 90% functional copy of the software. The two shell scripts it
calls on lines 222 and 227 will be included at a later date, as well as
a README on how to setup most of the easy-rsa interfacing parts. I'm
just committing it so Reimu doesn't fuck it up. Or maybe so he can fuck
it up. idk.